Split out settings and kobo types into separate packages
