+++ pkg/db/init.go
@@ -1,0 +1,1 @@
+package db
