+++ pkg/kobo/kobo.go
@@ -1,0 +1,1 @@
+package kobo
