+++ pkg/kobo/kobo_test.go
@@ -1,0 +1,1 @@
+package kobo
